handshake: validate HKDF-Expand-Label against crypto/tls implementation
